后台新增顶部进度条，优化ajax统一处理 增加产品索引操作

+++ mall-manager/src/main/java/me/liuhui/mall/manager/service/impl/AdSpaceServiceImpl.java
@@ -122,7 +122,17 @@ public class AdSpaceServiceImpl implements AdSpaceService {
     @Override
     public ResultVO<Boolean> delete(Set<Long> ids) {
         for (Long id : ids) {
+            AdSpace adSpace = adSpaceDao.selectByPk(id);
+            if (adSpace == null) {
+                return ResultVO.buildFailResult("广告位不存在！");
+            }
+            SaveAdHtmlDTO dto = new SaveAdHtmlDTO();
+            dto.setNo(adSpace.getNo());
+            dto.setHtml("");
+            fileService.saveAdHtml(dto);
+            adItemDao.deleteBySpaceId(id);
             adSpaceDao.delete(id);
+
         }
         return ResultVO.buildSuccessResult();
     }
